gsgpu: don't quit if GPIO request failed
During GSGPU backlight init we try to turn the LCD on. Some HW
(such as the TianChuangZhe NL430i laptop) don't seem to expose the
relevant pins (there is no public documentation anyway) so we
should ignore the errors if the GPIO pins cannot be requested.
The LCD most likely will have be turned on by the HW so it's ok
if the requests have seemingly failed.